dynamic noise model support added, user manuel updated with description of the new API
