+++ data/device/dummy_ibmq2.json
@@ -1,13 +1,69 @@
-{"name": "ibmq_dummy", "version": "0.1", "num_qubits": 2, "basis_gates": ["id", "rz", "sx", "x", "cx", "reset"], 
-    "T1": {"0": 2.8771109618032543e-06, "1": 1.3144669098122387e-06}, 
-    "T2": {"0": 3.4877983644780445e-06, "1": 7.152594920906184e-06}, 
-    "freq": {"0": 5193961135.028633, "1": 5105631508.014624}, 
-    "readout_length": {"0": 0.0, "1": 0.0}, 
-    "prob_meas0_prep1": {"0": 0.0, "1": 0.0}, 
-    "prob_meas1_prep0": {"0": 0.0, "1": 0.0}, 
-    "gate_lens": {"id0": 5.369133902008304e-09, "rz0": 0, "sx0": 8.667172242752573e-09, "x0": 0.0, "reset0": 7.1035042350812925e-09, "id1": 2.8152332051129814e-09, "rz1": 0, "sx1": 9.025154058657865e-09, "x1": 0.0, "reset1": 6.2512281658465e-09, "cx0_1": 0.0, "cx1_0": 0.0}, 
-    "gate_errs": {"id0": 5.317788598583956e-09, "rz0": 0, "sx0": 7.293775903821365e-09, 
-        "x0": 0.6, "reset0": 1.5610475657972145e-09, "id1": 3.6539622712438435e-09, "rz1": 0, "sx1": 3.2597073210988228e-09, 
-        "x1": 0.0, "reset1": 8.518604865879944e-09, 
-        "cx0_1": 0.0, 
-        "cx1_0": 0.0}, "cx_coupling": ["0_1", "1_0"]}
+{
+    "name": "ibmq_dummy",
+    "version": "0.1",
+    "num_qubits": 2,
+    "basis_gates": [
+        "id",
+        "rz",
+        "sx",
+        "x",
+        "cx",
+        "reset"
+    ],
+    "T1": {
+        "0": 1,
+        "1": 1
+    },
+    "T2": {
+        "0": 1,
+        "1": 1
+    },
+    "freq": {
+        "0": 1,
+        "1": 1
+    },
+    "readout_length": {
+        "0": 1,
+        "1": 1
+    },
+    "prob_meas0_prep1": {
+        "0": 1,
+        "1": 1
+    },
+    "prob_meas1_prep0": {
+        "0": 1,
+        "1": 1
+    },
+    "gate_lens": {
+        "id0": 1,
+        "rz0": 1,
+        "sx0": 1,
+        "x0": 1,
+        "reset0": 1,
+        "id1": 1,
+        "rz1": 1,
+        "sx1": 1,
+        "x1": 1,
+        "reset1": 1,
+        "cx0_1": 1,
+        "cx1_0": 1
+    },
+    "gate_errs": {
+        "id0": 1,
+        "rz0": 1,
+        "sx0": 1,
+        "x0": 1,
+        "reset0": 1,
+        "id1": 1,
+        "rz1": 1,
+        "sx1": 1,
+        "x1": 1,
+        "reset1": 1,
+        "cx0_1": 1,
+        "cx1_0": 1
+    },
+    "cx_coupling": [
+        "0_1",
+        "1_0"
+    ]
+}
